requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
cffi==1.11.5
configparser==3.5.0
cryptography==2.2.2
enum34==1.1.6
futures==3.2.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.18
pylint==1.9.2
pymongo==3.6.1
pyOpenSSL==18.0.0
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
